Refactor app.rb
