+++ books.json
@@ -1,0 +1,1 @@
+[{"title":"Book2","publisher":"Virag","cover_state":"bad","publish_date":"2020-09-08"},{"title":"Awesome books","publisher":"Microverse","cover_state":"good","publish_date":"2020-02-10"}]
